function/requirements.txt
autopep8==2.0.0
azure-functions==1.12.0
psycopg2==2.9.5
pycodestyle==2.9.1
python-http-client==3.3.7
sendgrid==6.9.7
starkbank-ecdsa==2.2.0
tomli==2.0.1
